Move crypto provider specific PEM reader code into separate classes.
